flatten orders and rxs
