+++ apps/app/src/views/routes/Order.tsx
@@ -1,5 +1,5 @@
 import { useParams } from 'react-router-dom';
-import { useState, useEffect } from 'react';
+import { useState, useEffect, useMemo } from 'react';
 import {
   Alert,
   AlertDescription,
@@ -21,9 +21,6 @@ import {
   Stack,
   Table,
   TableContainer,
-  Tag,
-  TagLabel,
-  TagLeftIcon,
   Tbody,
   Td,
   Text,
@@ -37,21 +34,12 @@ import {
 } from '@chakra-ui/react';
 import { gql, GraphQLClient } from 'graphql-request';
 import { usePhoton, types } from '@photonhealth/react';
-import {
-  FiAlertTriangle,
-  FiArrowUpRight,
-  FiCheck,
-  FiClock,
-  FiCopy,
-  FiCornerUpRight,
-  FiX,
-  FiChevronRight
-} from 'react-icons/fi';
+import { FiCopy, FiChevronRight } from 'react-icons/fi';
 import { Page } from '../components/Page';
 import PatientView from '../components/PatientView';
 import { confirmWrapper } from '../components/GuardDialog';
 import { formatAddress, formatDate, formatFills, formatPhone } from '../../utils';
-import { ORDER_FULFILLMENT_COLOR_MAP, ORDER_FULFILLMENT_STATE_MAP } from './Orders';
+import OrderStatusBadge, { OrderFulfillmentState } from '../components/OrderStatusBadge';
 export const graphQLClient = new GraphQLClient(process.env.REACT_APP_GRAPHQL_URI as string, {
   jsonSerializer: {
     parse: JSON.parse,
@@ -65,27 +53,11 @@ export const CANCEL_ORDER = gql`
     }
   }
 `;
-const ORDER_FULFILLMENT_TYPE_MAP = {
+export const ORDER_FULFILLMENT_TYPE_MAP = {
   [types.FulfillmentType.PickUp]: 'Pick up',
   [types.FulfillmentType.MailOrder]: 'Mail order'
 };
 
-export const ORDER_STATE_MAP: { [key in types.OrderState]: string } = {
-  PLACED: 'Placed',
-  ROUTING: 'Routing',
-  PENDING: 'Pending',
-  CANCELED: 'Canceled',
-  COMPLETED: 'Completed',
-  ERROR: 'Error'
-};
-export const ORDER_STATE_ICON_MAP: any = {
-  PLACED: FiArrowUpRight,
-  ROUTING: FiCornerUpRight,
-  PENDING: FiClock,
-  CANCELED: FiX,
-  COMPLETED: FiCheck,
-  ERROR: FiAlertTriangle
-};
 export const FILL_STATE_MAP: object = {
   CANCELED: 'Canceled',
   NEW: 'New',
@@ -98,6 +70,7 @@ export const FILL_COLOR_MAP: object = {
   SCHEDULED: 'orange',
   SENT: 'yellow'
 };
+
 export const Order = () => {
   const toast = useToast();
   const params = useParams();
@@ -211,6 +184,17 @@ export const Order = () => {
     </Button>
   );
 
+  const prescriptions = useMemo(() => {
+    if (!order) return [];
+
+    const rxIds = new Set();
+    return order.fills.filter((fill: any) => {
+      if (rxIds.has(fill.prescription.id)) return false;
+      rxIds.add(fill.prescription.id);
+      return true;
+    });
+  }, [order]);
+
   return (
     <Page header="Order" buttons={buttons}>
       <Card>
@@ -266,10 +250,10 @@ export const Order = () => {
                           ms={isMobile ? 'auto' : undefined}
                         />
                       ) : (
-                        <Tag size="sm" borderRadius="full">
-                          <TagLeftIcon boxSize="12px" as={ORDER_STATE_ICON_MAP[order.state]} />
-                          <TagLabel>{ORDER_STATE_MAP[order.state as keyof object] || ''}</TagLabel>
-                        </Tag>
+                        <OrderStatusBadge
+                          fulfillmentState={order.fulfillment?.state as OrderFulfillmentState}
+                          orderState={order.state}
+                        />
                       )}
                     </Td>
                   </Tr>
@@ -377,16 +361,10 @@ export const Order = () => {
                           ms={isMobile ? 'auto' : undefined}
                         />
                       ) : order.fulfillment?.state ? (
-                        <Badge
-                          size="sm"
-                          colorScheme={
-                            ORDER_FULFILLMENT_COLOR_MAP[order.fulfillment.state as keyof object] ||
-                            ''
-                          }
-                        >
-                          {ORDER_FULFILLMENT_STATE_MAP[order.fulfillment.state as keyof object] ||
-                            ''}
-                        </Badge>
+                        <OrderStatusBadge
+                          fulfillmentState={order.fulfillment?.state as OrderFulfillmentState}
+                          orderState={order.state}
+                        />
                       ) : null}
                     </Td>
                   </Tr>
@@ -520,9 +498,9 @@ export const Order = () => {
             <Text color="gray.500" fontWeight="medium" fontSize="sm">
               Fills
             </Text>
-            {order?.fills.length > 0 ? (
+            {prescriptions.length > 0 ? (
               <>
-                {order.fills.map((fill: any, i: number) => {
+                {prescriptions.map((fill: any, i: number) => {
                   return i < 5 ? (
                     <LinkBox key={fill.id} w="full" style={{ textDecoration: 'none' }}>
                       <Card
